moving test files for separate projects into own directories.
