+++ test/snapp_tests/snapp_auction.js
