+++ test/snapp_tests/snapp_utils.js
